+++ ajax/settings.php
@@ -10,51 +10,53 @@ npadmin_security(array("Administrators"), false);
 
 $returnList = false;
 
-if ($_POST['op'] == "add") {
-   $setting = new Setting($_POST);
-   if ($setting->store())
-      echo "OK";
-   else 
-      echo "ERROR";
-      
-} if ($_POST['op'] == "update") {
-   $setting = new Setting($_POST);
-   if ($setting->update())
-      echo "OK";
-   else 
-      echo "ERROR";
-      
-} else if ($_POST['op'] == "delete") {
-   $list = split(",", $_POST['list']);
-   foreach ($list as $id) {
-      $d = split("#", $id);
-      $setting = new Setting($d[1], $d[0]);
-      if (!$setting->delete()) {
-         echo "ERROR: Unable to delete setting '".$id[1]."'";
-         return;
+if (array_key_exists("op", $_POST)) {
+   if ($_POST['op'] == "add") {
+      $setting = new Setting($_POST);
+      if ($setting->store())
+         echo "OK";
+      else 
+         echo "ERROR";
+         
+   } if ($_POST['op'] == "update") {
+      $setting = new Setting($_POST);
+      if ($setting->update())
+         echo "OK";
+      else 
+         echo "ERROR";
+         
+   } else if ($_POST['op'] == "delete") {
+      $list = split(",", $_POST['list']);
+      foreach ($list as $id) {
+         $d = split("#", $id);
+         $setting = new Setting($d[1], $d[0]);
+         if (!$setting->delete()) {
+            echo "ERROR: Unable to delete setting '".$id[1]."'";
+            return;
+         }
       }
-   }
-   echo "OK";
-   
-} else if ($_POST['op'] == "list" || isset($_GET['op']) && $_GET['op'] == "list") {
-   $returnList = true;
-}
-
-if ($returnList) {
-   $settings = array();
-
-   function createSettingsList($setting) {
-      global $settings;
-      $settings[] = $setting;
-   }
-
-   if (in_array('type', array_keys($_POST)) && isset($_POST['type']) && $_POST['type'] != "ALL")
-      $sql = "SELECT * FROM ".$ddbb->getTable("Setting")." WHERE ".$ddbb->getMapping('Setting','type')." = ".NP_DDBB::encodeSQLValue($_POST['type'], $ddbb->getType('Setting','type'))." ORDER BY type, name";
-   else 
-      $sql = "SELECT * FROM ".$ddbb->getTable("Setting")." ORDER BY type, name";
+      echo "OK";
       
-   $ddbb->executeSelectQuery($sql, "createSettingsList");
-
-   echo json_encode($settings); 
-} 
+   } else if ($_POST['op'] == "list" || isset($_GET['op']) && $_GET['op'] == "list") {
+      $returnList = true;
+   }
+   
+   if ($returnList) {
+      $settings = array();
+   
+      function createSettingsList($setting) {
+         global $settings;
+         $settings[] = $setting;
+      }
+   
+      if (in_array('type', array_keys($_POST)) && isset($_POST['type']) && $_POST['type'] != "ALL")
+         $sql = "SELECT * FROM ".$ddbb->getTable("Setting")." WHERE ".$ddbb->getMapping('Setting','type')." = ".NP_DDBB::encodeSQLValue($_POST['type'], $ddbb->getType('Setting','type'))." ORDER BY type, name";
+      else 
+         $sql = "SELECT * FROM ".$ddbb->getTable("Setting")." ORDER BY type, name";
+         
+      $ddbb->executeSelectQuery($sql, "createSettingsList");
+   
+      echo json_encode($settings); 
+   } 
+}
 ?>
